+++ components/core/core-tenants/src/main/java/org/eclipse/dirigible/components/tenants/service/TenantService.java
@@ -14,7 +14,8 @@ import java.util.Optional;
 import java.util.Set;
 import org.eclipse.dirigible.components.tenants.domain.Tenant;
 import org.eclipse.dirigible.components.tenants.domain.TenantStatus;
-import org.eclipse.dirigible.components.tenants.tenant.TenantContextInitFilter;
+import org.eclipse.dirigible.components.tenants.tenant.TenantExtractor;
+import org.eclipse.dirigible.components.tenants.tenant.TenantImpl;
 import org.springframework.stereotype.Service;
 
 /**
@@ -72,7 +73,7 @@ public class TenantService {
      */
     public Tenant save(Tenant tenant) {
         Tenant newTenant = tenantRepository.save(tenant);
-        TenantContextInitFilter.TENANT_CACHE.invalidateAll();
+        TenantExtractor.TENANT_CACHE.put(newTenant.getSubdomain(), Optional.of(TenantImpl.createFromEntity(newTenant)));
         return newTenant;
     }
 
@@ -93,7 +94,7 @@ public class TenantService {
      */
     public void delete(Tenant tenant) {
         tenantRepository.delete(tenant);
-        TenantContextInitFilter.TENANT_CACHE.invalidateAll();
+        TenantExtractor.TENANT_CACHE.invalidate(tenant.getSubdomain());
     }
 
 }
