+++ components/core/core-tenants/src/main/java/org/eclipse/dirigible/components/tenants/tenant/TenantContextInitFilter.java
@@ -10,29 +10,16 @@
 package org.eclipse.dirigible.components.tenants.tenant;
 
 import java.io.IOException;
-import java.security.Principal;
-import java.util.Arrays;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
-import java.util.concurrent.TimeUnit;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-import org.eclipse.dirigible.commons.config.DirigibleConfig;
 import org.eclipse.dirigible.components.base.tenant.Tenant;
 import org.eclipse.dirigible.components.base.tenant.TenantContext;
-import org.eclipse.dirigible.components.tenants.service.TenantService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
-import com.github.benmanes.caffeine.cache.Cache;
-import com.github.benmanes.caffeine.cache.Caffeine;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import jakarta.servlet.FilterChain;
@@ -51,21 +38,11 @@ public class TenantContextInitFilter extends OncePerRequestFilter {
 
     private static final List<String> HOST_HEADERS = List.of("host", "x-forwarded-host");
 
-    /** The Constant TENANT_SUBDOMAIN_PATTERN. */
-    private static final Pattern TENANT_SUBDOMAIN_PATTERN = Pattern.compile(DirigibleConfig.TENANT_SUBDOMAIN_REGEX.getStringValue());
-
-    /** The Constant tenantCache. */
-    public static final Cache<String, Optional<Tenant>> TENANT_CACHE = Caffeine.newBuilder()
-                                                                               .expireAfterWrite(10, TimeUnit.MINUTES)
-                                                                               .maximumSize(100)
-                                                                               .build();
-
     /** The tenant service. */
-    private final TenantService tenantService;
+    private final TenantExtractor tenantExtractor;
 
     /** The tenant context. */
     private final TenantContext tenantContext;
-    private final boolean multitenantModeEnabled;
     private final Gson gson;
 
     /**
@@ -74,10 +51,9 @@ public class TenantContextInitFilter extends OncePerRequestFilter {
      * @param tenantService the tenant service
      * @param tenantContext the tenant context
      */
-    public TenantContextInitFilter(TenantService tenantService, TenantContext tenantContext) {
-        this.tenantService = tenantService;
+    public TenantContextInitFilter(TenantExtractor tenantExtractor, TenantContext tenantContext) {
+        this.tenantExtractor = tenantExtractor;
         this.tenantContext = tenantContext;
-        this.multitenantModeEnabled = DirigibleConfig.MULTI_TENANT_MODE_ENABLED.getBooleanValue();
         this.gson = new GsonBuilder().serializeNulls()
                                      .create();
     }
@@ -94,39 +70,13 @@ public class TenantContextInitFilter extends OncePerRequestFilter {
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain)
             throws ServletException, IOException {
-        Optional<Tenant> currentTenant = determineTenantSubdomain(request);
+        Optional<Tenant> currentTenant = tenantExtractor.determineTenantSubdomain(request);
         if (currentTenant.isEmpty()) {
             response.sendError(HttpServletResponse.SC_NOT_FOUND, "There is no registered tenant for the current host");
             if (LOGGER.isWarnEnabled()) {
                 LOGGER.warn("Tried to reach unregistered tenant. Headers: [{}]", getHeaders(request));
             }
             return;
-        }
-
-        if (multitenantModeEnabled) {
-            Principal principal = request.getUserPrincipal();
-            if (principal instanceof OAuth2AuthenticationToken oauthToken) {
-                String tenantAttribute = oauthToken.getPrincipal()
-                                                   .getAttribute("custom:tenant");
-                if (tenantAttribute == null || tenantAttribute.equals("")) {
-                    forbidden("User is not assigned to any tenant", response);
-                    return;
-                }
-                Set<String> tenants = new HashSet<>(Arrays.asList(tenantAttribute.split(","))
-                                                          .stream()
-                                                          .map(e -> e.trim())
-                                                          .collect(Collectors.toList()));
-                Optional<Tenant> tenant = determineTenantSubdomain(request);
-                if (tenant.isEmpty() || !tenants.contains(tenant.get()
-                                                                .getSubdomain())) {
-                    forbidden("User is not member of the [" + tenant.get()
-                                                                    .getName()
-                            + " | " + tenant.get()
-                                            .getSubdomain()
-                            + "] tenant", response);
-                    return;
-                }
-            }
         }
 
         try {
@@ -142,82 +92,10 @@ public class TenantContextInitFilter extends OncePerRequestFilter {
         }
     }
 
-    /**
-     * Determine current tenant.
-     *
-     * @param request the request
-     * @return the optional
-     */
-    private Optional<Tenant> determineTenantSubdomain(HttpServletRequest request) {
-        if (!multitenantModeEnabled) {
-            LOGGER.debug("The app is in single tenant mode. Will return the default tenant.");
-            return Optional.of(TenantImpl.getDefaultTenant());
-        }
-        List<Optional<String>> subdomains = HOST_HEADERS.stream()
-                                                        .map(hostHeader -> determineTenantSubdomain(request, hostHeader))
-                                                        .toList();
-        if (areEmpty(subdomains)) {
-            if (LOGGER.isDebugEnabled()) {
-                LOGGER.debug("Tenant subdomain cannot be extracted from the current request. Will return the default tenant. Headers: [{}]",
-                        getHeaders(request));
-            }
-            return Optional.of(TenantImpl.getDefaultTenant());
-        }
-
-        for (Optional<String> subdomain : subdomains) {
-            Optional<Tenant> tenant = tenantBySubdomain(subdomain);
-            if (tenant.isPresent()) {
-                LOGGER.debug("Found registered tenant [{}] for subdomain [{}].", tenant.get(), subdomain.get());
-                return tenant;
-            }
-        }
-        return Optional.empty();
-    }
-
     private String getHeaders(HttpServletRequest request) {
         Map<String, String> headers = new HashMap<>();
         HOST_HEADERS.forEach(h -> headers.put(h, request.getHeader(h)));
         return gson.toJson(headers);
-    }
-
-    private boolean areEmpty(List<Optional<String>> optionals) {
-        for (Optional<?> optional : optionals) {
-            if (optional.isPresent()) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    private Optional<String> determineTenantSubdomain(HttpServletRequest request, String hostHeader) {
-        String headerValue = request.getHeader(hostHeader);
-        if (null == headerValue) {
-            LOGGER.debug("Cannot extract subdomain from header [{}] with null value.", hostHeader);
-            return Optional.empty();
-        }
-        Matcher matcher = TENANT_SUBDOMAIN_PATTERN.matcher(headerValue);
-        if (matcher.find()) {
-            String tenantSubdomain = matcher.group(1);
-            LOGGER.debug("Host [{}] from header [{}] MATCHES tenant subdomain pattern [{}]. Extracted subdomain: [{}]", headerValue,
-                    hostHeader, TENANT_SUBDOMAIN_PATTERN.pattern(), tenantSubdomain);
-
-            return Optional.of(tenantSubdomain);
-        }
-        LOGGER.debug("Host [{}] from header [{}] does NOT match tenant subdomain pattern [{}].", headerValue, hostHeader,
-                TENANT_SUBDOMAIN_PATTERN.pattern());
-        return Optional.empty();
-    }
-
-    private Optional<Tenant> tenantBySubdomain(Optional<String> subdomainOpt) {
-        if (subdomainOpt.isEmpty()) {
-            return Optional.empty();
-        }
-        String subdomain = subdomainOpt.get();
-        return TENANT_CACHE.get(subdomain, k -> {
-            LOGGER.debug("Searching for tenant with subdomain [{}] from database", subdomain);
-            return tenantService.findBySubdomain(subdomain)
-                                .map(TenantImpl::createFromEntity);
-        });
     }
 
     /**
@@ -235,26 +113,7 @@ public class TenantContextInitFilter extends OncePerRequestFilter {
                 || request.getRequestURI()
                           .startsWith("/js/")
                 || request.getRequestURI()
-                          .endsWith(".ico")
-                || request.getRequestURI()
-                          .startsWith("/index-busy.html")
-                || request.getRequestURI()
-                          .startsWith("/services/js/platform-core/extension-services/themes.js")
-                || request.getRequestURI()
-                          .startsWith("/services/js/platform-core/services/loader.js");
+                          .endsWith(".ico");
     }
 
-    /**
-     * Forbidden.
-     *
-     * @param message the message
-     * @param response the response
-     * @throws IOException Signals that an I/O exception has occurred.
-     */
-    private void forbidden(String message, HttpServletResponse response) throws IOException {
-        if (logger.isWarnEnabled()) {
-            logger.warn(message);
-        }
-        response.sendError(HttpServletResponse.SC_FORBIDDEN, message);
-    }
 }
