+++ components/pom.xml
@@ -34,6 +34,7 @@
 		<module>security/security-keycloak</module>
 		<module>security/security-oauth2</module>
 		<module>security/security-snowflake</module>
+		<module>security/security-cognito</module>
 
 		<!-- Data -->
 		<module>data/data-structures</module>
@@ -479,6 +480,11 @@
 			<dependency>
 				<groupId>org.eclipse.dirigible</groupId>
 				<artifactId>dirigible-components-security-snowflake</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>org.eclipse.dirigible</groupId>
+				<artifactId>dirigible-components-security-cognito</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 			<dependency>
