+++ components/security/security-cognito/src/main/java/org/eclipse/dirigible/components/security/cognito/CognitoSecurityConfiguration.java
@@ -7,23 +7,31 @@
  *
  * SPDX-FileCopyrightText: Eclipse Dirigible contributors SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.dirigible.components.security.oauth2;
+package org.eclipse.dirigible.components.security.cognito;
 
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 import org.eclipse.dirigible.components.base.http.access.HttpSecurityURIConfigurator;
+import org.eclipse.dirigible.components.base.util.AuthoritiesUtil;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.mapping.GrantedAuthoritiesMapper;
+import org.springframework.security.oauth2.core.oidc.user.OidcUserAuthority;
 import org.springframework.security.web.SecurityFilterChain;
 
 /**
  * The Class OAuth2SecurityConfiguration.
  */
-@Profile("github")
+@Profile("cognito")
 @Configuration
-public class OAuth2SecurityConfiguration {
+public class CognitoSecurityConfiguration {
 
     /**
      * Filter chain.
@@ -34,18 +42,33 @@ public class OAuth2SecurityConfiguration {
      */
     @Bean
     SecurityFilterChain filterChain(HttpSecurity http, HttpSecurityURIConfigurator httpSecurityURIConfigurator) throws Exception {
-        http//
-            .authorizeHttpRequests(authz -> authz.requestMatchers("/oauth2/**", "/login/**")
+        http.authorizeHttpRequests(authz -> authz.requestMatchers("/oauth2/**", "/login/**")
                                                  .permitAll())
             .csrf(csrf -> csrf.disable())
             .headers(headers -> headers.frameOptions(frameOpts -> frameOpts.disable()))
             .oauth2Client(Customizer.withDefaults())
-            .oauth2Login(Customizer.withDefaults())
+            .oauth2Login(oauth2 -> {
+                oauth2.userInfoEndpoint(userInfoEndpointConfig -> userInfoEndpointConfig.userAuthoritiesMapper(userAuthoritiesMapper()));
+            })
             .oauth2ResourceServer(oauth2 -> oauth2.jwt(Customizer.withDefaults()))
             .sessionManagement(sessionManagement -> sessionManagement.sessionCreationPolicy(SessionCreationPolicy.ALWAYS));
 
         httpSecurityURIConfigurator.configure(http);
 
         return http.build();
+    }
+
+    @Bean
+    public GrantedAuthoritiesMapper userAuthoritiesMapper() {
+        return (authorities) -> {
+            OidcUserAuthority oidcUserAuthority = (OidcUserAuthority) new ArrayList<>(authorities).get(0);
+            List<String> cognitoGroups = (ArrayList<String>) oidcUserAuthority.getAttributes()
+                                                                              .get("cognito:groups");
+            Set<GrantedAuthority> grantedAuthorities = new HashSet<>();
+            if (cognitoGroups != null) {
+                grantedAuthorities.addAll(AuthoritiesUtil.toAuthorities(cognitoGroups));
+            }
+            return grantedAuthorities;
+        };
     }
 }
