+++ components/security/security-cognito/src/main/resources/application-cognito.properties
