+++ modules/commons/commons-config/src/main/java/org/eclipse/dirigible/commons/config/DirigibleConfig.java
@@ -9,11 +9,10 @@
  */
 package org.eclipse.dirigible.commons.config;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import java.nio.charset.StandardCharsets;
 import java.util.Base64;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * The Enum DirigibleConfig.
@@ -58,6 +57,9 @@ public enum DirigibleConfig {
 
     /** The multi tenant mode enabled. */
     MULTI_TENANT_MODE_ENABLED("DIRIGIBLE_MULTI_TENANT_MODE", Boolean.FALSE.toString()),
+
+    /** The multi tenant mode single user pool enabled. */
+    MULTI_TENANT_MODE_SINGLE_USER_POOL("DIRIGIBLE_MULTI_TENANT_MODE_SINGLE_USER_POOL", Boolean.FALSE.toString()),
 
     /** The tenant subdomain regex. */
     TENANT_SUBDOMAIN_REGEX("DIRIGIBLE_TENANT_SUBDOMAIN_REGEX", "^([^\\.]+)\\..+$"),
